+++ custom_components/econet300/api.py
@@ -1,48 +1,66 @@
+"""Module provides the API functionality for ecoNET-300 Home Assistant Integration."""
 import asyncio
-import logging
 from http import HTTPStatus
+import logging
 from typing import Any
 
-from aiohttp import ClientSession, BasicAuth
+from aiohttp import BasicAuth, ClientSession
+
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
 
 from .const import (
+    API_EDITABLE_PARAMS_LIMITS_DATA,
+    API_EDITABLE_PARAMS_LIMITS_URI,
+    API_REG_PARAMS_PARAM_DATA,
+    API_REG_PARAMS_URI,
+    API_SYS_PARAMS_PARAM_HW_VER,
+    API_SYS_PARAMS_PARAM_SW_REV,
     API_SYS_PARAMS_PARAM_UID,
     API_SYS_PARAMS_URI,
-    API_REG_PARAMS_URI,
-    API_REG_PARAMS_PARAM_DATA,
-    API_SYS_PARAMS_PARAM_SW_REV,
-    API_SYS_PARAMS_PARAM_HW_VER,
+    EDITABLE_PARAMS_MAPPING_TABLE,
 )
 from .mem_cache import MemCache
 
 _LOGGER = logging.getLogger(__name__)
 
 
+def map_param(param_name):
+    """Check params mapping in const.py."""
+    if param_name not in EDITABLE_PARAMS_MAPPING_TABLE:
+        return None
+
+    return EDITABLE_PARAMS_MAPPING_TABLE[param_name]
+
+
 class Limits:
+    """Class representing the limits of a value."""
+
     def __init__(self, min_v: float, max_v: float):
         self.min = min_v
         self.max = max_v
 
-
+    class AuthError(Exception):
+        """Raised when authentication fails."""
 class AuthError(Exception):
-    """AuthError"""
+    """Raised when authentication fails."""
 
 
 class ApiError(Exception):
-    """AuthError"""
+    """Raised when an API error occurs."""
 
 
 class DataError(Exception):
-    """DataError"""
+    """Raised when there is an error with the data."""
 
 
 class EconetClient:
+    """Client for interacting with the ecoNET-300 API."""
+
     def __init__(
         self, host: str, username: str, password: str, session: ClientSession
     ) -> None:
-        """Initialize."""
+        """Initializethe EconetClient."""
 
         proto = ["http://", "https://"]
 
@@ -57,17 +75,25 @@ class EconetClient:
         self._auth = BasicAuth(username, password)
 
     def host(self):
+        """Get the host."""
         return self._host
 
     async def set_param(self, key: str, value: str):
-        url = "{}/econet/rmCurrNewParam?newParamKey={}&newParamValue={}".format(
-            self._host, key, value
-        )
+        """Set a parameter."""
+        url = f"{self._host}/econet/rmCurrNewParam?newParamKey={key}&newParamValue={value}"
 
         return await self._get(url)
 
     async def get_params(self, reg: str):
-        url = "{}/econet/{}".format(self._host, reg)
+        """Get parameters for a given registry.
+
+        Args:
+            reg (str): The registry to retrieve parameters from.
+
+        Returns:
+            dict: The parameters retrieved from the registry.
+        """
+        url = f"{self._host}/econet/{reg}"
 
         return await self._get(url)
 
@@ -96,6 +122,8 @@ class EconetClient:
 
 
 class Econet300Api:
+    """Client for interacting with the ecoNET-300 API."""
+
     def __init__(self, client: EconetClient, cache: MemCache) -> None:
         self._client = client
         self._cache = cache
@@ -105,24 +133,30 @@ class Econet300Api:
 
     @classmethod
     async def create(cls, client: EconetClient, cache: MemCache):
+        """Create an instance of Econet300Api."""
         c = cls(client, cache)
         await c.init()
 
         return c
 
     def host(self):
+        """Get the host."""
         return self._client.host()
 
     def uid(self) -> str:
+        """Get the UID."""
         return self._uid
 
     def sw_rev(self) -> str:
-        return self._sw_revision
+        """Get the software revision."""
+        # Set a parameter value via the Econet 300 API.
 
     def hw_ver(self) -> str:
+        """Get the hardware version."""
         return self._hw_version
 
     async def init(self):
+        """Initialize the Econet300Api."""
         sys_params = await self._client.get_params(API_SYS_PARAMS_URI)
 
         if API_SYS_PARAMS_PARAM_UID not in sys_params:
@@ -150,6 +184,7 @@ class Econet300Api:
             self._hw_version = sys_params[API_SYS_PARAMS_PARAM_HW_VER]
 
     async def set_param(self, param, value) -> bool:
+        """Set a parameter value via the Econet 300 API."""
         param_idx = map_param(param)
         if param_idx is None:
             _LOGGER.warning(
@@ -170,6 +205,7 @@ class Econet300Api:
         return True
 
     async def get_param_limits(self, param: str):
+        """Fetch and return the limits for a particular parameter from the Econet 300 API, using a cache for efficient retrieval if available."""
         if not self._cache.exists(API_EDITABLE_PARAMS_LIMITS_DATA):
             limits = await self._fetch_reg_key(
                 API_EDITABLE_PARAMS_LIMITS_URI, API_EDITABLE_PARAMS_LIMITS_DATA
@@ -206,9 +242,7 @@ class Econet300Api:
         return {**reg_params, **sys_params}
 
     async def _fetch_reg_key(self, reg, data_key: str | None = None):
-        """Fetch a key from the json-encoded data.
-        If key is None, then return whole data.
-        """
+        """Fetch a key from the json-encoded data returned by the API for a given registry If key is None, then return whole data."""
         data = await self._client.get_params(reg)
 
         if data is None:
@@ -225,6 +259,7 @@ class Econet300Api:
 
 
 async def make_api(hass: HomeAssistant, cache: MemCache, data: dict):
+    """Create an Econet 300 API instance."""
     return await Econet300Api.create(
         EconetClient(
             data["host"],